Refactor onion circuit boundaries
